microservices/notification-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.7

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Email providers
sendgrid==6.10.0

# SMS providers  
twilio==8.10.0

# Environment variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Production server
gunicorn==21.2.0

# Logging
structlog==23.2.0
